chore: refactor after rebase for side drawer
