+++ packages/frontend/src/components/MultiRowResultVariables/utils.tsx
